Fixes to "staying on the same page functionality"
The new implementation should be very flexible.

Fixes #7226

Additionally, this allows using the version switcher even if the website is published at a different URL than the `site_url`. In particular, the version switcher should now work locally when serving the site with `mike serve`.

The tests should demonstrate correctness of the implementation.